+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/tutorial/ClientUI.java
@@ -61,7 +61,7 @@ public final class ClientUI {
         // insert the entity
         editModel.insert();
         // clear the edit model after a successful insert
-        editModel.setDefaults();
+        editModel.defaults();
       }
       catch (DatabaseException | ValidationException e) {
         JOptionPane.showMessageDialog((JTextField) actionEvent.getSource(),
@@ -125,7 +125,7 @@ public final class ClientUI {
         // insert the entity
         editModel.insert();
         // clear the edit model after a successful insert
-        editModel.setDefaults();
+        editModel.defaults();
         // and transfer the focus to the combo box
         artistComboBox.requestFocusInWindow();
       }