+++ framework/model-test/src/main/java/is/codion/framework/model/test/AbstractEntityModelTest.java
@@ -148,7 +148,7 @@ public abstract class AbstractEntityModelTest<Model extends DefaultEntityModel<M
     departmentModel.tableModel().selectionModel().setSelectedIndexes(asList(1, 2, 3));
     assertTrue(departmentModel.tableModel().selectionModel().selectionNotEmpty().get());
     assertTrue(departmentModel.editModel().exists().get());
-    departmentModel.editModel().setDefaults();
+    departmentModel.editModel().defaults();
     assertTrue(departmentModel.tableModel().selectionModel().selectionEmpty().get());
   }