+++ swing/framework-ui/src/test/java/is/codion/swing/framework/ui/component/EntityComponentsTest.java
@@ -68,7 +68,7 @@ public final class EntityComponentsTest {
 
   @Test
   void checkBox() {
-    editModel.setDefaults();
+    editModel.defaults();
     ComponentValue<Boolean, JCheckBox> componentValue =
             entityComponents.checkBox(Detail.BOOLEAN)
                     .transferFocusOnEnter(true)
@@ -89,7 +89,7 @@ public final class EntityComponentsTest {
 
   @Test
   void toggleButton() {
-    editModel.setDefaults();
+    editModel.defaults();
     ComponentValue<Boolean, JToggleButton> componentValue =
             entityComponents.toggleButton(Detail.BOOLEAN)
                     .transferFocusOnEnter(true)
@@ -110,7 +110,7 @@ public final class EntityComponentsTest {
 
   @Test
   void nullableCheckBox() {
-    editModel.setDefaults();
+    editModel.defaults();
     ComponentValue<Boolean, JCheckBox> componentValue =
             entityComponents.checkBox(Detail.BOOLEAN_NULLABLE)
                     .transferFocusOnEnter(true)
@@ -132,7 +132,7 @@ public final class EntityComponentsTest {
 
   @Test
   void booleanComboBox() {
-    editModel.setDefaults();
+    editModel.defaults();
     editModel.put(Detail.BOOLEAN, true);
     ComponentValue<Boolean, JComboBox<Item<Boolean>>> componentValue =
             entityComponents.booleanComboBox(Detail.BOOLEAN)